requirements.txt
discord.py==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.3
beautifulsoup4==4.12.2
discord.py==2.3.2
flask==2.3.3
requests==2.31.0